#157 * Completed migration of fcp.App to be at the top level. Where previously you would specify ‘fcp.app()…’, you now simply specify ‘fcp…’ * NOTE: Two main changes to how you call common functions: 	1. All functions are now methods. Where previously you would call ‘fcp.launch()’, you now call ‘fcp:launch()’ instead. 	2. Functions which returned a value, such as ‘fcp.running()’ have been renamed to ‘isXXX’ or ‘getXXX’ for consistency. So, ‘fcp.running()’ is now ‘fcp:isRunning()’
