+++ src/hs/fcpxhacks/modules/fcpx10-3.lua
@@ -266,11 +266,11 @@ function loadScript()
 	--------------------------------------------------------------------------------
 	local lastFinalCutProVersion = settings.get("fcpxHacks.lastFinalCutProVersion")
 	if lastFinalCutProVersion == nil then
-		settings.set("fcpxHacks.lastFinalCutProVersion", fcp.app():getVersion())
+		settings.set("fcpxHacks.lastFinalCutProVersion", fcp:getVersion())
 	else
-		if lastFinalCutProVersion ~= fcp.app():getVersion() then
+		if lastFinalCutProVersion ~= fcp:getVersion() then
 			settings.set("fcpxHacks.chooserMenuItems", nil) -- Reset Chooser Menu Items.
-			settings.set("fcpxHacks.lastFinalCutProVersion", fcp.app():getVersion())
+			settings.set("fcpxHacks.lastFinalCutProVersion", fcp:getVersion())
 		end
 	end
 
@@ -292,11 +292,11 @@ function loadScript()
 	else
 		if tonumber(settings.get("fcpxHacks.lastVersion")) < tonumber(fcpxhacks.scriptVersion) then
 			if settings.get("fcpxHacks.enableHacksShortcutsInFinalCutPro") then
-				local finalCutProRunning = fcp.app():isRunning()
+				local finalCutProRunning = fcp:isRunning()
 				if finalCutProRunning then
 					dialog.displayMessage(i18n("newKeyboardShortcuts"))
 					updateKeyboardShortcuts()
-					if not fcp.fcp.app():restart() then
+					if not fcp.fcp:restart() then
 						--------------------------------------------------------------------------------
 						-- Failed to restart Final Cut Pro:
 						--------------------------------------------------------------------------------
@@ -441,7 +441,7 @@ function loadScript()
 	--------------------------------------------------------------------------------
 	-- Bind Keyboard Shortcuts:
 	--------------------------------------------------------------------------------
-	mod.lastCommandSet = fcp.app():getActiveCommandSetPath()
+	mod.lastCommandSet = fcp:getActiveCommandSetPath()
 	bindKeyboardShortcuts()
 
 	--------------------------------------------------------------------------------
@@ -454,7 +454,7 @@ function loadScript()
 	--------------------------------------------------------------------------------
 	-- Activate the correct modal state:
 	--------------------------------------------------------------------------------
-	if fcp.app():isFrontmost() then
+	if fcp:isFrontmost() then
 
 		--------------------------------------------------------------------------------
 		-- Used by Watchers to prevent double-ups:
@@ -894,7 +894,7 @@ function bindKeyboardShortcuts()
 		--------------------------------------------------------------------------------
 		-- Update Active Command Set:
 		--------------------------------------------------------------------------------
-		fcp.app():getActiveCommandSet(nil, true)
+		fcp:getActiveCommandSet(nil, true)
 
 		--------------------------------------------------------------------------------
 		-- Use Default Shortcuts Keys:
@@ -965,7 +965,7 @@ end
 --------------------------------------------------------------------------------
 function getShortcutsFromActiveCommandSet()
 
-	local activeCommandSetTable = fcp.app():getActiveCommandSet(nil, true)
+	local activeCommandSetTable = fcp:getActiveCommandSet(nil, true)
 
 	if activeCommandSetTable ~= nil then
 		for k, v in pairs(mod.finalCutProShortcutKeyPlaceholders) do
@@ -1085,7 +1085,7 @@ function updateKeyboardShortcuts()
 	--------------------------------------------------------------------------------
 	-- Revert back to default keyboard layout:
 	--------------------------------------------------------------------------------
-	local result = fcp.app():setPreference("Active Command Set", fcp.app():getPath() .. "/Contents/Resources/" .. fcp.currentLanguage() .. ".lproj/Default.commandset")
+	local result = fcp:setPreference("Active Command Set", fcp:getPath() .. "/Contents/Resources/" .. fcp.currentLanguage() .. ".lproj/Default.commandset")
 	if not result then
 		dialog.displayErrorMessage(i18n("activeCommandSetResetError"))
 		return false
@@ -1098,8 +1098,8 @@ end
 --------------------------------------------------------------------------------
 function enableHacksShortcuts()
 	local appleScript = [[
-		set finalCutProPath to "]] .. fcp.app():getPath() .. [["
-		set finalCutProLanguages to ]] .. inspect(fcp.app():getSupportedLanguages()) .. [[
+		set finalCutProPath to "]] .. fcp:getPath() .. [["
+		set finalCutProLanguages to ]] .. inspect(fcp:getSupportedLanguages()) .. [[
 
 		--------------------------------------------------------------------------------
 		-- Replace Files:
@@ -1142,8 +1142,8 @@ end
 --------------------------------------------------------------------------------
 function disableHacksShortcuts()
 	local appleScript = [[
-		set finalCutProPath to "]] .. fcp.app():getPath() .. [["
-		set finalCutProLanguages to ]] .. inspect(fcp.app():getSupportedLanguages()) .. [[
+		set finalCutProPath to "]] .. fcp:getPath() .. [["
+		set finalCutProLanguages to ]] .. inspect(fcp:getSupportedLanguages()) .. [[
 
 		try
 			do shell script "cp -f ~/.hammerspoon/hs/fcpxhacks/plist/10-3/old/NSProCommandGroups.plist '" & finalCutProPath & "/Contents/Resources/NSProCommandGroups.plist'" with administrator privileges
@@ -1208,8 +1208,8 @@ end
 		--------------------------------------------------------------------------------
 		-- Assume FCPX is closed if not told otherwise:
 		--------------------------------------------------------------------------------
-		local fcpxActive = fcp.app():isFrontmost()
-		local fcpxRunning = fcp.app():isRunning()
+		local fcpxActive = fcp:isFrontmost()
+		local fcpxRunning = fcp:isRunning()
 
 		--------------------------------------------------------------------------------
 		-- We only refresh plist values if necessary as this takes time:
@@ -1224,7 +1224,7 @@ end
 			--------------------------------------------------------------------------------
 			-- Read Final Cut Pro Preferences:
 			--------------------------------------------------------------------------------
-			local preferences = fcp.app():getPreferences()
+			local preferences = fcp:getPreferences()
 			if preferences == nil then
 				dialog.displayErrorMessage(i18n("failedToReadFCPPreferences"))
 				return "Fail"
@@ -1235,7 +1235,7 @@ end
 			--------------------------------------------------------------------------------
 			mod.allowMovingMarkers = false
 
-			local result = plist.fileToTable(fcp.app():getPath() .. "/Contents/Frameworks/TLKit.framework/Versions/A/Resources/EventDescriptions.plist")
+			local result = plist.fileToTable(fcp:getPath() .. "/Contents/Frameworks/TLKit.framework/Versions/A/Resources/EventDescriptions.plist")
 			if result ~= nil then
 				if result["TLKMarkerHandler"] ~= nil then
 					if result["TLKMarkerHandler"]["Configuration"] ~= nil then
@@ -1497,7 +1497,7 @@ end
 						if file:sub(-7) == ".fcpxml" then
 							emptySharedXMLFiles = false
 							local xmlPath = xmlSharingPath .. folder .. "/" .. file
-							table.insert(submenu, {title = file:sub(1, -8), fn = function() fcp.app():importXML(xmlPath) end, disabled = not fcpxRunning})
+							table.insert(submenu, {title = file:sub(1, -8), fn = function() fcp:importXML(xmlPath) end, disabled = not fcpxRunning})
 						end
 					end
 
@@ -1536,7 +1536,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Current Language:
 		--------------------------------------------------------------------------------
-		local currentLanguage = fcp.app():getCurrentLanguage()
+		local currentLanguage = fcp:getCurrentLanguage()
 
 		--------------------------------------------------------------------------------
 		-- Effects Shortcuts:
@@ -1786,11 +1786,11 @@ end
 		}
 		local shortcutsTable = {
 			{ title = string.upper(i18n("shortcuts")) .. ":", 											disabled = true },
-			{ title = i18n("createOptimizedMedia"), 													fn = function() toggleCreateOptimizedMedia() end, 					checked = fcp.app():getPreference("FFImportCreateOptimizeMedia", false),				disabled = not fcpxRunning },
-			{ title = i18n("createMulticamOptimizedMedia"),												fn = function() toggleCreateMulticamOptimizedMedia() end, 			checked = fcp.app():getPreference("FFCreateOptimizedMediaForMulticamClips", true), 	disabled = not fcpxRunning },
-			{ title = i18n("createProxyMedia"), 														fn = function() toggleCreateProxyMedia() end, 						checked = fcp.app():getPreference("FFImportCreateProxyMedia", false),					disabled = not fcpxRunning },
-			{ title = i18n("leaveFilesInPlaceOnImport"), 												fn = function() toggleLeaveInPlace() end, 							checked = not fcp.app():getPreference("FFImportCopyToMediaFolder", true),				disabled = not fcpxRunning },
-			{ title = i18n("enableBackgroundRender").." ("..mod.FFAutoRenderDelay.." "..i18n("secs")..")", fn = function() toggleBackgroundRender() end, 						checked = fcp.app():getPreference("FFAutoStartBGRender", true),						disabled = not fcpxRunning },
+			{ title = i18n("createOptimizedMedia"), 													fn = function() toggleCreateOptimizedMedia() end, 					checked = fcp:getPreference("FFImportCreateOptimizeMedia", false),				disabled = not fcpxRunning },
+			{ title = i18n("createMulticamOptimizedMedia"),												fn = function() toggleCreateMulticamOptimizedMedia() end, 			checked = fcp:getPreference("FFCreateOptimizedMediaForMulticamClips", true), 	disabled = not fcpxRunning },
+			{ title = i18n("createProxyMedia"), 														fn = function() toggleCreateProxyMedia() end, 						checked = fcp:getPreference("FFImportCreateProxyMedia", false),					disabled = not fcpxRunning },
+			{ title = i18n("leaveFilesInPlaceOnImport"), 												fn = function() toggleLeaveInPlace() end, 							checked = not fcp:getPreference("FFImportCopyToMediaFolder", true),				disabled = not fcpxRunning },
+			{ title = i18n("enableBackgroundRender").." ("..mod.FFAutoRenderDelay.." "..i18n("secs")..")", fn = function() toggleBackgroundRender() end, 						checked = fcp:getPreference("FFAutoStartBGRender", true),						disabled = not fcpxRunning },
 			{ title = "-" },
 		}
 		local automationOptions = {
@@ -1887,7 +1887,7 @@ end
 		local proxyMenuIcon = ""
 
 		local proxyStatusIcon = nil
-		local FFPlayerQuality = fcp.app():getPreference("FFPlayerQuality")
+		local FFPlayerQuality = fcp:getPreference("FFPlayerQuality")
 		if FFPlayerQuality == 4 then
 			proxyStatusIcon = "🔴" 		-- Proxy (4)
 		else
@@ -1939,8 +1939,8 @@ end
 		if enableHacksShortcutsInFinalCutPro == nil then enableHacksShortcutsInFinalCutPro = false end
 
 		if enableHacksShortcutsInFinalCutPro then
-			if fcp.app():isRunning() then
-				fcp.app():launch()
+			if fcp:isRunning() then
+				fcp:launch()
 				fcp:app():commandEditor():show()
 			end
 		else
@@ -2007,7 +2007,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure Final Cut Pro is active:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
+		fcp:launch()
 
 		--------------------------------------------------------------------------------
 		-- Warning message:
@@ -2017,13 +2017,13 @@ end
 		--------------------------------------------------------------------------------
 		-- Save the layout of the Transitions panel in case we switch away...
 		--------------------------------------------------------------------------------
-		local transitions = fcp.app():transitions()
+		local transitions = fcp:transitions()
 		local transitionsLayout = transitions:saveLayout()
 
 		--------------------------------------------------------------------------------
 		-- Make sure Effects panel is open:
 		--------------------------------------------------------------------------------
-		local effects = fcp.app():effects()
+		local effects = fcp:effects()
 		local effectsShowing = effects:isShowing()
 		if not effects:show():isShowing() then
 			dialog.displayErrorMessage("Unable to activate the Effects panel.\n\nError occurred in updateEffectsList().")
@@ -2139,7 +2139,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure Final Cut Pro is active:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
+		fcp:launch()
 
 		--------------------------------------------------------------------------------
 		-- Warning message:
@@ -2149,7 +2149,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Save the layout of the Effects panel, in case we switch away...
 		--------------------------------------------------------------------------------
-		local effects = fcp.app():effects()
+		local effects = fcp:effects()
 		local effectsLayout = nil
 		if effects:isShowing() then
 			effectsLayout = effects:saveLayout()
@@ -2158,7 +2158,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure Transitions panel is open:
 		--------------------------------------------------------------------------------
-		local transitions = fcp.app():transitions()
+		local transitions = fcp:transitions()
 		local transitionsShowing = transitions:isShowing()
 		if not transitions:show():isShowing() then
 			dialog.displayErrorMessage("Unable to activate the Transitions panel.\n\nError occurred in updateEffectsList().")
@@ -2241,7 +2241,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure Final Cut Pro is active:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
+		fcp:launch()
 
 		--------------------------------------------------------------------------------
 		-- Hide the Touch Bar:
@@ -2253,7 +2253,7 @@ end
 		--------------------------------------------------------------------------------
 		dialog.displayMessage(i18n("updateTitlesListWarning"))
 
-		local app = fcp.app()
+		local app = fcp
 		local generators = app:generators()
 
 		local browserLayout = app:browser():saveLayout()
@@ -2334,7 +2334,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure Final Cut Pro is active:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
+		fcp:launch()
 
 		--------------------------------------------------------------------------------
 		-- Hide the Touch Bar:
@@ -2346,7 +2346,7 @@ end
 		--------------------------------------------------------------------------------
 		dialog.displayMessage(i18n("updateGeneratorsListWarning"))
 
-		local app = fcp.app()
+		local app = fcp
 		local generators = app:generators()
 
 		local browserLayout = app:browser():saveLayout()
@@ -2429,7 +2429,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Was Final Cut Pro Open?
 		--------------------------------------------------------------------------------
-		mod.wasFinalCutProOpen = fcp.app():isFrontmost()
+		mod.wasFinalCutProOpen = fcp:isFrontmost()
 
 		--------------------------------------------------------------------------------
 		-- Get settings:
@@ -2548,7 +2548,7 @@ end
 			--------------------------------------------------------------------------------
 			-- Put focus back in Final Cut Pro:
 			--------------------------------------------------------------------------------
-			if mod.wasFinalCutProOpen then fcp.app():launch() end
+			if mod.wasFinalCutProOpen then fcp:launch() end
 
 			--------------------------------------------------------------------------------
 			-- Refresh Menubar:
@@ -2565,7 +2565,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Was Final Cut Pro Open?
 		--------------------------------------------------------------------------------
-		mod.wasFinalCutProOpen = fcp.app():isFrontmost()
+		mod.wasFinalCutProOpen = fcp:isFrontmost()
 
 		--------------------------------------------------------------------------------
 		-- Get settings:
@@ -2665,7 +2665,7 @@ end
 			--------------------------------------------------------------------------------
 			-- Put focus back in Final Cut Pro:
 			--------------------------------------------------------------------------------
-			if mod.wasFinalCutProOpen then fcp.app():launch() end
+			if mod.wasFinalCutProOpen then fcp:launch() end
 
 			--------------------------------------------------------------------------------
 			-- Refresh Menubar:
@@ -2682,7 +2682,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Was Final Cut Pro Open?
 		--------------------------------------------------------------------------------
-		mod.wasFinalCutProOpen = fcp.app():isFrontmost()
+		mod.wasFinalCutProOpen = fcp:isFrontmost()
 
 		--------------------------------------------------------------------------------
 		-- Get settings:
@@ -2782,7 +2782,7 @@ end
 			--------------------------------------------------------------------------------
 			-- Put focus back in Final Cut Pro:
 			--------------------------------------------------------------------------------
-			if mod.wasFinalCutProOpen then fcp.app():launch() end
+			if mod.wasFinalCutProOpen then fcp:launch() end
 
 			--------------------------------------------------------------------------------
 			-- Refresh Menubar:
@@ -2799,7 +2799,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Was Final Cut Pro Open?
 		--------------------------------------------------------------------------------
-		mod.wasFinalCutProOpen = fcp.app():isFrontmost()
+		mod.wasFinalCutProOpen = fcp:isFrontmost()
 
 		--------------------------------------------------------------------------------
 		-- Get settings:
@@ -2899,7 +2899,7 @@ end
 			--------------------------------------------------------------------------------
 			-- Put focus back in Final Cut Pro:
 			--------------------------------------------------------------------------------
-			if mod.wasFinalCutProOpen then fcp.app():launch() end
+			if mod.wasFinalCutProOpen then fcp:launch() end
 
 			--------------------------------------------------------------------------------
 			-- Refresh Menubar:
@@ -2916,7 +2916,7 @@ end
 	-- CHANGE BATCH EXPORT DESTINATION PRESET:
 	--------------------------------------------------------------------------------
 	function changeBatchExportDestinationPreset()
-		local shareMenuItems = fcp.app():menuBar():findMenuItemsUI("File", "Share")
+		local shareMenuItems = fcp:menuBar():findMenuItemsUI("File", "Share")
 		if not shareMenuItems then
 			dialog.displayErrorMessage(i18n("batchExportDestinationsNotFound"))
 			return
@@ -2970,7 +2970,7 @@ end
 		-- If Final Cut Pro is running...
 		--------------------------------------------------------------------------------
 		local restartStatus = false
-		if fcp.app():isRunning() then
+		if fcp:isRunning() then
 			if dialog.displayYesNoQuestion(i18n("changeFinalCutProLanguage") .. "\n\n" .. i18n("doYouWantToContinue")) then
 				restartStatus = true
 			else
@@ -2981,7 +2981,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Update Final Cut Pro's settings::
 		--------------------------------------------------------------------------------
-		local result = fcp.app():setPreference("AppleLanguages", {language})
+		local result = fcp:setPreference("AppleLanguages", {language})
 		if not result then
 			dialog.displayErrorMessage(i18n("failedToChangeLanguage"))
 		end
@@ -2989,13 +2989,13 @@ end
 		--------------------------------------------------------------------------------
 		-- Change FCPX Hacks Language:
 		--------------------------------------------------------------------------------
-		fcp.app():getCurrentLanguage(true, language)
+		fcp:getCurrentLanguage(true, language)
 
 		--------------------------------------------------------------------------------
 		-- Restart Final Cut Pro:
 		--------------------------------------------------------------------------------
 		if restartStatus then
-			if not fcp.fcp.app():restart() then
+			if not fcp.fcp:restart() then
 				--------------------------------------------------------------------------------
 				-- Failed to restart Final Cut Pro:
 				--------------------------------------------------------------------------------
@@ -3055,17 +3055,17 @@ end
 		--------------------------------------------------------------------------------
 		-- Get existing value:
 		--------------------------------------------------------------------------------
-		if fcp.app():getPreference("FFPeriodicBackupInterval") == nil then
+		if fcp:getPreference("FFPeriodicBackupInterval") == nil then
 			mod.FFPeriodicBackupInterval = 15
 		else
-			mod.FFPeriodicBackupInterval = fcp.app():getPreference("FFPeriodicBackupInterval")
+			mod.FFPeriodicBackupInterval = fcp:getPreference("FFPeriodicBackupInterval")
 		end
 
 		--------------------------------------------------------------------------------
 		-- If Final Cut Pro is running...
 		--------------------------------------------------------------------------------
 		local restartStatus = false
-		if fcp.app():isRunning() then
+		if fcp:isRunning() then
 			if dialog.displayYesNoQuestion(i18n("changeBackupInterval") .. "\n\n" .. doYouWantToContinue) then
 				restartStatus = true
 			else
@@ -3084,7 +3084,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Update plist:
 		--------------------------------------------------------------------------------
-		local result = fcp.app():setPreference("FFPeriodicBackupInterval", tostring(userSelectedBackupInterval))
+		local result = fcp:setPreference("FFPeriodicBackupInterval", tostring(userSelectedBackupInterval))
 		if result == nil then
 			dialog.displayErrorMessage(i18n("backupIntervalFail"))
 			return "Failed"
@@ -3099,7 +3099,7 @@ end
 		-- Restart Final Cut Pro:
 		--------------------------------------------------------------------------------
 		if restartStatus then
-			if not fcp.fcp.app():restart() then
+			if not fcp.fcp:restart() then
 				--------------------------------------------------------------------------------
 				-- Failed to restart Final Cut Pro:
 				--------------------------------------------------------------------------------
@@ -3123,14 +3123,14 @@ end
 		--------------------------------------------------------------------------------
 		-- Get existing value:
 		--------------------------------------------------------------------------------
-		local executeResult,executeStatus = execute("/usr/libexec/PlistBuddy -c \"Print :FFOrganizerSmartCollections\" '" .. fcp.app():getPath() .. "/Contents/Frameworks/Flexo.framework/Versions/A/Resources/en.lproj/FFLocalizable.strings'")
+		local executeResult,executeStatus = execute("/usr/libexec/PlistBuddy -c \"Print :FFOrganizerSmartCollections\" '" .. fcp:getPath() .. "/Contents/Frameworks/Flexo.framework/Versions/A/Resources/en.lproj/FFLocalizable.strings'")
 		if tools.trim(executeResult) ~= "" then FFOrganizerSmartCollections = executeResult end
 
 		--------------------------------------------------------------------------------
 		-- If Final Cut Pro is running...
 		--------------------------------------------------------------------------------
 		local restartStatus = false
-		if fcp.app():isRunning() then
+		if fcp:isRunning() then
 			if dialog.displayYesNoQuestion(i18n("changeSmartCollectionsLabel") .. "\n\n" .. i18n("doYouWantToContinue")) then
 				restartStatus = true
 			else
@@ -3150,8 +3150,8 @@ end
 		-- Update plist for every Flexo language:
 		--------------------------------------------------------------------------------
 		local executeCommands = {}
-		for k, v in pairs(fcp.app():getFlexoLanguages()) do
-			local executeCommand = "/usr/libexec/PlistBuddy -c \"Set :FFOrganizerSmartCollections " .. tools.trim(userSelectedSmartCollectionsLabel) .. "\" '" .. fcp.app():getPath() .. "/Contents/Frameworks/Flexo.framework/Versions/A/Resources/" .. fcp.app():getFlexoLanguages()[k] .. ".lproj/FFLocalizable.strings'"
+		for k, v in pairs(fcp:getFlexoLanguages()) do
+			local executeCommand = "/usr/libexec/PlistBuddy -c \"Set :FFOrganizerSmartCollections " .. tools.trim(userSelectedSmartCollectionsLabel) .. "\" '" .. fcp:getPath() .. "/Contents/Frameworks/Flexo.framework/Versions/A/Resources/" .. fcp:getFlexoLanguages()[k] .. ".lproj/FFLocalizable.strings'"
 			executeCommands[#executeCommands + 1] = executeCommand
 		end
 		local result = tools.executeWithAdministratorPrivileges(executeCommands)
@@ -3164,7 +3164,7 @@ end
 		-- Restart Final Cut Pro:
 		--------------------------------------------------------------------------------
 		if restartStatus then
-			if not fcp.fcp.app():restart() then
+			if not fcp.fcp:restart() then
 				--------------------------------------------------------------------------------
 				-- Failed to restart Final Cut Pro:
 				--------------------------------------------------------------------------------
@@ -3216,7 +3216,7 @@ end
 			-- Stop Watchers:
 			--------------------------------------------------------------------------------
 			mod.scrollingTimelineWatcherDown:stop()
-			fcp.app():timeline():unlockPlayhead()
+			fcp:timeline():unlockPlayhead()
 
 			--------------------------------------------------------------------------------
 			-- Display Notification:
@@ -3271,8 +3271,8 @@ end
 		local lockTimelinePlayhead = settings.get("fcpxHacks.lockTimelinePlayhead") or false
 
 		if lockTimelinePlayhead then
-			if fcp.app():isRunning() then
-				fcp.app():timeline():unlockPlayhead()
+			if fcp:isRunning() then
+				fcp:timeline():unlockPlayhead()
 			end
 			dialog.displayNotification(i18n("playheadLockDeactivated"))
 			settings.set("fcpxHacks.lockTimelinePlayhead", false)
@@ -3286,8 +3286,8 @@ end
 				toggleScrollingTimeline()
 				message = i18n("scrollingTimelineDeactivated") .. "\n"
 			end
-			if fcp.app():isRunning() then
-				fcp.app():timeline():lockPlayhead()
+			if fcp:isRunning() then
+				fcp:timeline():lockPlayhead()
 			end
 			dialog.displayNotification(message..i18n("playheadLockActivated"))
 			settings.set("fcpxHacks.lockTimelinePlayhead", true)
@@ -3323,7 +3323,7 @@ end
 				settings.set("fcpxHacks.enableVoiceCommands", enableVoiceCommands)
 				return
 			end
-			if fcp.app():isFrontmost() then
+			if fcp:isFrontmost() then
 				voicecommands:start()
 			else
 				voicecommands:stop()
@@ -3343,7 +3343,7 @@ end
 		if enableHacksHUD then
 			hackshud.hide()
 		else
-			if fcp.app():isFrontmost() then
+			if fcp:isFrontmost() then
 				hackshud.show()
 			end
 		end
@@ -3512,7 +3512,7 @@ end
 
 		if not enableMobileNotifications then
 
-			local returnToFinalCutPro = fcp.app():isFrontmost()
+			local returnToFinalCutPro = fcp:isFrontmost()
 			::retryProwlAPIKeyEntry::
 
 			local result = dialog.displayTextBoxMessage(i18n("mobileNotificationsTextbox"), i18n("mobileNotificationsError") .. "\n\n" .. i18n("pleaseTryAgain"), prowlAPIKey)
@@ -3522,7 +3522,7 @@ end
 			end
 			local prowlAPIKeyValidResult, prowlAPIKeyValidError = prowlAPIKeyValid(result)
 			if prowlAPIKeyValidResult then
-				if returnToFinalCutPro then fcp.app():launch() end
+				if returnToFinalCutPro then fcp:launch() end
 				settings.set("fcpxHacks.prowlAPIKey", result)
 				notificationWatcher()
 				settings.set("fcpxHacks.enableMobileNotifications", not enableMobileNotifications)
@@ -3616,7 +3616,7 @@ end
 		-- If Final Cut Pro is running...
 		--------------------------------------------------------------------------------
 		local restartStatus = false
-		if fcp.app():isRunning() then
+		if fcp:isRunning() then
 			if dialog.displayYesNoQuestion(enableOrDisableText .. " " .. i18n("hacksShortcutsRestart") .. " " .. i18n("doYouWantToContinue")) then
 				restartStatus = true
 			else
@@ -3636,7 +3636,7 @@ end
 			--------------------------------------------------------------------------------
 			-- Revert back to default keyboard layout:
 			--------------------------------------------------------------------------------
-			local result = fcp.app():setPreference("Active Command Set", fcp.app():getPath() .. "/Contents/Resources/en.lproj/Default.commandset")
+			local result = fcp:setPreference("Active Command Set", fcp:getPath() .. "/Contents/Resources/en.lproj/Default.commandset")
 			if result == nil then
 				dialog.displayErrorMessage(i18n("activeCommandSetResetError"))
 				return "Failed"
@@ -3654,7 +3654,7 @@ end
 			--------------------------------------------------------------------------------
 			-- Revert back to default keyboard layout:
 			--------------------------------------------------------------------------------
-			local result = fcp.app():setPreference("Active Command Set", fcp.app():getPath() .. "/Contents/Resources/en.lproj/Default.commandset")
+			local result = fcp:setPreference("Active Command Set", fcp:getPath() .. "/Contents/Resources/en.lproj/Default.commandset")
 			if result == nil then
 				dialog.displayErrorMessage(i18n("activeCommandSetResetError"))
 				return "Failed"
@@ -3680,7 +3680,7 @@ end
 		-- Restart Final Cut Pro:
 		--------------------------------------------------------------------------------
 		if restartStatus then
-			if not fcp.fcp.app():restart() then
+			if not fcp.fcp:restart() then
 				--------------------------------------------------------------------------------
 				-- Failed to restart Final Cut Pro:
 				--------------------------------------------------------------------------------
@@ -3736,14 +3736,14 @@ end
 		-- Get existing value:
 		--------------------------------------------------------------------------------
 		mod.allowMovingMarkers = false
-		local executeResult,executeStatus = execute("/usr/libexec/PlistBuddy -c \"Print :TLKMarkerHandler:Configuration:'Allow Moving Markers'\" '" .. fcp.app():getPath() .. "/Contents/Frameworks/TLKit.framework/Versions/A/Resources/EventDescriptions.plist'")
+		local executeResult,executeStatus = execute("/usr/libexec/PlistBuddy -c \"Print :TLKMarkerHandler:Configuration:'Allow Moving Markers'\" '" .. fcp:getPath() .. "/Contents/Frameworks/TLKit.framework/Versions/A/Resources/EventDescriptions.plist'")
 		if tools.trim(executeResult) == "true" then mod.allowMovingMarkers = true end
 
 		--------------------------------------------------------------------------------
 		-- If Final Cut Pro is running...
 		--------------------------------------------------------------------------------
 		local restartStatus = false
-		if fcp.app():isRunning() then
+		if fcp:isRunning() then
 			if dialog.displayYesNoQuestion(i18n("togglingMovingMarkersRestart") .. "\n\n" .. i18n("doYouWantToContinue")) then
 				restartStatus = true
 			else
@@ -3755,13 +3755,13 @@ end
 		-- Update plist:
 		--------------------------------------------------------------------------------
 		if mod.allowMovingMarkers then
-			local executeStatus = tools.executeWithAdministratorPrivileges([[/usr/libexec/PlistBuddy -c \"Set :TLKMarkerHandler:Configuration:'Allow Moving Markers' false\" ']] .. fcp.app():getPath() .. [[/Contents/Frameworks/TLKit.framework/Versions/A/Resources/EventDescriptions.plist']])
+			local executeStatus = tools.executeWithAdministratorPrivileges([[/usr/libexec/PlistBuddy -c \"Set :TLKMarkerHandler:Configuration:'Allow Moving Markers' false\" ']] .. fcp:getPath() .. [[/Contents/Frameworks/TLKit.framework/Versions/A/Resources/EventDescriptions.plist']])
 			if executeStatus == false then
 				dialog.displayErrorMessage(i18n("movingMarkersError"))
 				return "Failed"
 			end
 		else
-			local executeStatus = tools.executeWithAdministratorPrivileges([[/usr/libexec/PlistBuddy -c \"Set :TLKMarkerHandler:Configuration:'Allow Moving Markers' true\" ']] .. fcp.app():getPath() .. [[/Contents/Frameworks/TLKit.framework/Versions/A/Resources/EventDescriptions.plist']])
+			local executeStatus = tools.executeWithAdministratorPrivileges([[/usr/libexec/PlistBuddy -c \"Set :TLKMarkerHandler:Configuration:'Allow Moving Markers' true\" ']] .. fcp:getPath() .. [[/Contents/Frameworks/TLKit.framework/Versions/A/Resources/EventDescriptions.plist']])
 			if executeStatus == false then
 				dialog.displayErrorMessage(i18n("movingMarkersError"))
 				return "Failed"
@@ -3772,7 +3772,7 @@ end
 		-- Restart Final Cut Pro:
 		--------------------------------------------------------------------------------
 		if restartStatus then
-			if not fcp.fcp.app():restart() then
+			if not fcp.fcp:restart() then
 				--------------------------------------------------------------------------------
 				-- Failed to restart Final Cut Pro:
 				--------------------------------------------------------------------------------
@@ -3801,17 +3801,17 @@ end
 		--------------------------------------------------------------------------------
 		-- Get existing value:
 		--------------------------------------------------------------------------------
-		if fcp.app():getPreference("FFSuspendBGOpsDuringPlay") == nil then
+		if fcp:getPreference("FFSuspendBGOpsDuringPlay") == nil then
 			mod.FFSuspendBGOpsDuringPlay = false
 		else
-			mod.FFSuspendBGOpsDuringPlay = fcp.app():getPreference("FFSuspendBGOpsDuringPlay")
+			mod.FFSuspendBGOpsDuringPlay = fcp:getPreference("FFSuspendBGOpsDuringPlay")
 		end
 
 		--------------------------------------------------------------------------------
 		-- If Final Cut Pro is running...
 		--------------------------------------------------------------------------------
 		local restartStatus = false
-		if fcp.app():isRunning() then
+		if fcp:isRunning() then
 			if dialog.displayYesNoQuestion(i18n("togglingBackgroundTasksRestart") .. "\n\n" ..i18n("doYouWantToContinue")) then
 				restartStatus = true
 			else
@@ -3822,7 +3822,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Update plist:
 		--------------------------------------------------------------------------------
-		local result = fcp.app():setPreference("FFSuspendBGOpsDuringPlay", mod.FFSuspendBGOpsDuringPlay == true)
+		local result = fcp:setPreference("FFSuspendBGOpsDuringPlay", mod.FFSuspendBGOpsDuringPlay == true)
 		if result == nil then
 			dialog.displayErrorMessage(i18n("failedToWriteToPreferences"))
 			return "Failed"
@@ -3832,7 +3832,7 @@ end
 		-- Restart Final Cut Pro:
 		--------------------------------------------------------------------------------
 		if restartStatus then
-			if not fcp.fcp.app():restart() then
+			if not fcp.fcp:restart() then
 				--------------------------------------------------------------------------------
 				-- Failed to restart Final Cut Pro:
 				--------------------------------------------------------------------------------
@@ -3861,17 +3861,17 @@ end
 		--------------------------------------------------------------------------------
 		-- Get existing value:
 		--------------------------------------------------------------------------------
-		if fcp.app():getPreference("FFEnableGuards") == nil then
+		if fcp:getPreference("FFEnableGuards") == nil then
 			mod.FFEnableGuards = false
 		else
-			mod.FFEnableGuards = fcp.app():getPreference("FFEnableGuards")
+			mod.FFEnableGuards = fcp:getPreference("FFEnableGuards")
 		end
 
 		--------------------------------------------------------------------------------
 		-- If Final Cut Pro is running...
 		--------------------------------------------------------------------------------
 		local restartStatus = false
-		if fcp.app():isRunning() then
+		if fcp:isRunning() then
 			if dialog.displayYesNoQuestion(i18n("togglingTimecodeOverlayRestart") .. "\n\n" .. i18n("doYouWantToContinue")) then
 				restartStatus = true
 			else
@@ -3882,7 +3882,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Update plist:
 		--------------------------------------------------------------------------------
-		local result = fcp.app():setPreference("FFEnableGuards", mod.FFEnableGuards == true)
+		local result = fcp:setPreference("FFEnableGuards", mod.FFEnableGuards == true)
 		if result == nil then
 			dialog.displayErrorMessage(i18n("failedToWriteToPreferences"))
 			return "Failed"
@@ -3892,7 +3892,7 @@ end
 		-- Restart Final Cut Pro:
 		--------------------------------------------------------------------------------
 		if restartStatus then
-			if not fcp.fcp.app():restart() then
+			if not fcp.fcp:restart() then
 				--------------------------------------------------------------------------------
 				-- Failed to restart Final Cut Pro:
 				--------------------------------------------------------------------------------
@@ -3939,13 +3939,13 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure it's active:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
+		fcp:launch()
 
 		--------------------------------------------------------------------------------
 		-- If we're setting rather than toggling...
 		--------------------------------------------------------------------------------
 		log.d("optionalValue: "..inspect(optionalValue))
-		if optionalValue ~= nil and optionalValue == fcp.app():getPreference("FFCreateOptimizedMediaForMulticamClips", true) then
+		if optionalValue ~= nil and optionalValue == fcp:getPreference("FFCreateOptimizedMediaForMulticamClips", true) then
 			log.d("optionalValue matches preference value. Bailing.")
 			return
 		end
@@ -3977,12 +3977,12 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure it's active:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
+		fcp:launch()
 
 		--------------------------------------------------------------------------------
 		-- If we're setting rather than toggling...
 		--------------------------------------------------------------------------------
-		if optionalValue ~= nil and optionalValue == fcp.app():getPreference("FFImportCreateProxyMedia", false) then
+		if optionalValue ~= nil and optionalValue == fcp:getPreference("FFImportCreateProxyMedia", false) then
 			return
 		end
 
@@ -4013,12 +4013,12 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure it's active:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
+		fcp:launch()
 
 		--------------------------------------------------------------------------------
 		-- If we're setting rather than toggling...
 		--------------------------------------------------------------------------------
-		if optionalValue ~= nil and optionalValue == fcp.app():getPreference("FFImportCreateOptimizeMedia", false) then
+		if optionalValue ~= nil and optionalValue == fcp:getPreference("FFImportCreateOptimizeMedia", false) then
 			return
 		end
 
@@ -4050,12 +4050,12 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure it's active:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
+		fcp:launch()
 
 		--------------------------------------------------------------------------------
 		-- If we're setting rather than toggling...
 		--------------------------------------------------------------------------------
-		if optionalValue ~= nil and optionalValue == fcp.app():getPreference("FFImportCopyToMediaFolder", true) then
+		if optionalValue ~= nil and optionalValue == fcp:getPreference("FFImportCopyToMediaFolder", true) then
 			return
 		end
 
@@ -4087,12 +4087,12 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure it's active:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
+		fcp:launch()
 
 		--------------------------------------------------------------------------------
 		-- If we're setting rather than toggling...
 		--------------------------------------------------------------------------------
-		if optionalValue ~= nil and optionalValue == fcp.app():getPreference("FFAutoStartBGRender", true) then
+		if optionalValue ~= nil and optionalValue == fcp:getPreference("FFAutoStartBGRender", true) then
 			return
 		end
 
@@ -4129,14 +4129,14 @@ end
 		-- Write data back to Clipboard:
 		--------------------------------------------------------------------------------
 		clipboard.stopWatching()
-		pasteboard.writeDataForUTI(fcp.app():getPasteboardUTI(), data)
+		pasteboard.writeDataForUTI(fcp:getPasteboardUTI(), data)
 		clipboard.startWatching()
 
 		--------------------------------------------------------------------------------
 		-- Paste in FCPX:
 		--------------------------------------------------------------------------------
-		fcp.app():launch()
-		if not fcp.app():performShortcut("Paste") then
+		fcp:launch()
+		if not fcp:performShortcut("Paste") then
 			dialog.displayErrorMessage("Failed to trigger the 'Paste' Shortcut.\n\nError occurred in finalCutProPasteFromClipboardHistory().")
 			return "Failed"
 		end
@@ -4161,14 +4161,14 @@ end
 				-- Write data back to Clipboard:
 				--------------------------------------------------------------------------------
 				clipboard.stopWatching()
-				pasteboard.writeDataForUTI(fcp.app():getPasteboardUTI(), currentClipboardData)
+				pasteboard.writeDataForUTI(fcp:getPasteboardUTI(), currentClipboardData)
 				clipboard.startWatching()
 
 				--------------------------------------------------------------------------------
 				-- Paste in FCPX:
 				--------------------------------------------------------------------------------
-				fcp.app():launch()
-				if not fcp.app():performShortcut("Paste") then
+				fcp:launch()
+				if not fcp:performShortcut("Paste") then
 					dialog.displayErrorMessage("Failed to trigger the 'Paste' Shortcut.\n\nError occurred in pasteFromSharedClipboard().")
 					return "Failed"
 				end
@@ -4251,7 +4251,7 @@ end
 	--------------------------------------------------------------------------------
 	function resetSettings()
 
-		local finalCutProRunning = fcp.app():isRunning()
+		local finalCutProRunning = fcp:isRunning()
 
 		local resetMessage = i18n("trashFCPXHacksPreferences")
 		if finalCutProRunning then
@@ -4286,7 +4286,7 @@ end
 		-- Restart Final Cut Pro if running:
 		--------------------------------------------------------------------------------
 		if finalCutProRunning then
-			if not fcp.fcp.app():restart() then
+			if not fcp.fcp:restart() then
 				--------------------------------------------------------------------------------
 				-- Failed to restart Final Cut Pro:
 				--------------------------------------------------------------------------------
@@ -4345,7 +4345,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Check to see if the Keyword Editor is already open:
 		--------------------------------------------------------------------------------
-		local fcpx = fcp.app():application()
+		local fcpx = fcp:application()
 		local fcpxElements = ax.applicationElement(fcpx)
 		local whichWindow = nil
 		for i=1, fcpxElements:attributeValueCount("AXChildren") do
@@ -4466,7 +4466,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Check to see if the Keyword Editor is already open:
 		--------------------------------------------------------------------------------
-		local fcpx = fcp.app():application()
+		local fcpx = fcp:application()
 		local fcpxElements = ax.applicationElement(fcpx)
 		local whichWindow = nil
 		for i=1, fcpxElements:attributeValueCount("AXChildren") do
@@ -4877,7 +4877,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Check the option is available in the current context
 		--------------------------------------------------------------------------------
-		if not fcp.app():menuBar():isEnabled("File", "Reveal in Browser") then
+		if not fcp:menuBar():isEnabled("File", "Reveal in Browser") then
 			return nil
 		end
 		
@@ -4886,7 +4886,7 @@ end
 		--------------------------------------------------------------------------------
 		deleteAllHighlights()
 
-		local libraries = fcp.app():libraries()
+		local libraries = fcp:libraries()
 		local selectedClips
 		
 		
@@ -4898,7 +4898,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Trigger the menu item to reveal the clip
 		--------------------------------------------------------------------------------
-		fcp.app():menuBar():selectMenu("File", "Reveal in Browser")
+		fcp:menuBar():selectMenu("File", "Reveal in Browser")
 
 		--------------------------------------------------------------------------------
 		-- Give FCPX time to find the clip
@@ -4928,7 +4928,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Get Clip Name from the Viewer
 		--------------------------------------------------------------------------------
-		local clipName = fcp.app():viewer():getTitle()
+		local clipName = fcp:viewer():getTitle()
 		
 		if clipName then
 			--------------------------------------------------------------------------------
@@ -4969,7 +4969,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Show the Color Board with the correct panel
 		--------------------------------------------------------------------------------
-		local colorBoard = fcp.app():colorBoard()
+		local colorBoard = fcp:colorBoard()
 
 		--------------------------------------------------------------------------------
 		-- Show the Color Board if it's hidden:
@@ -5089,13 +5089,13 @@ end
 		--------------------------------------------------------------------------------
 		-- Save the Effects Browser layout:
 		--------------------------------------------------------------------------------
-		local effects = fcp.app():effects()
+		local effects = fcp:effects()
 		local effectsLayout = effects:saveLayout()
 
 		--------------------------------------------------------------------------------
 		-- Get Transitions Browser:
 		--------------------------------------------------------------------------------
-		local transitions = fcp.app():transitions()
+		local transitions = fcp:transitions()
 		local transitionsShowing = transitions:isShowing()
 		local transitionsLayout = transitions:saveLayout()
 
@@ -5198,13 +5198,13 @@ end
 		--------------------------------------------------------------------------------
 		-- Save the Transitions Browser layout:
 		--------------------------------------------------------------------------------
-		local transitions = fcp.app():transitions()
+		local transitions = fcp:transitions()
 		local transitionsLayout = transitions:saveLayout()
 
 		--------------------------------------------------------------------------------
 		-- Get Effects Browser:
 		--------------------------------------------------------------------------------
-		local effects = fcp.app():effects()
+		local effects = fcp:effects()
 		local effectsShowing = effects:isShowing()
 		local effectsLayout = effects:saveLayout()
 
@@ -5308,13 +5308,13 @@ end
 		--------------------------------------------------------------------------------
 		-- Save the main Browser layout:
 		--------------------------------------------------------------------------------
-		local browser = fcp.app():browser()
+		local browser = fcp:browser()
 		local browserLayout = browser:saveLayout()
 
 		--------------------------------------------------------------------------------
 		-- Get Titles Browser:
 		--------------------------------------------------------------------------------
-		local generators = fcp.app():generators()
+		local generators = fcp:generators()
 		local generatorsShowing = generators:isShowing()
 		local generatorsLayout = generators:saveLayout()
 
@@ -5423,13 +5423,13 @@ end
 		--------------------------------------------------------------------------------
 		-- Save the main Browser layout:
 		--------------------------------------------------------------------------------
-		local browser = fcp.app():browser()
+		local browser = fcp:browser()
 		local browserLayout = browser:saveLayout()
 
 		--------------------------------------------------------------------------------
 		-- Get Titles Browser:
 		--------------------------------------------------------------------------------
-		local generators = fcp.app():generators()
+		local generators = fcp:generators()
 		local generatorsShowing = generators:isShowing()
 		local generatorsLayout = generators:saveLayout()
 
@@ -5536,7 +5536,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Find the Timeline Appearance Button:
 		--------------------------------------------------------------------------------
-		local appearance = fcp.app():timeline():toolbar():appearance()
+		local appearance = fcp:timeline():toolbar():appearance()
 		appearance:show()
 		if direction == "up" then
 			appearance:clipHeight():increment()
@@ -5550,7 +5550,7 @@ end
 	--------------------------------------------------------------------------------
 	function changeTimelineClipHeightRelease()
 		mod.changeTimelineClipHeightAlreadyInProgress = false
-		fcp.app():timeline():toolbar():appearance():hide()
+		fcp:timeline():toolbar():appearance():hide()
 	end
 
 	--------------------------------------------------------------------------------
@@ -5589,7 +5589,7 @@ end
 	--------------------------------------------------------------------------------
 	function menuItemShortcut(i, x, y, z)
 
-		local fcpxElements = ax.applicationElement(fcp.app():application())
+		local fcpxElements = ax.applicationElement(fcp:application())
 
 		local whichMenuBar = nil
 		for i=1, fcpxElements:attributeValueCount("AXChildren") do
@@ -5635,7 +5635,7 @@ end
 		-- Toggle Touch Bar:
 		--------------------------------------------------------------------------------
 		setTouchBarLocation()
-		if fcp.app():isRunning() then
+		if fcp:isRunning() then
 			mod.touchBarWindow:toggle()
 		end
 
@@ -5657,27 +5657,27 @@ end
 	function cutAndSwitchMulticam(whichMode, whichAngle)
 
 		if whichMode == "Audio" then
-			if not fcp.app():performShortcut("MultiAngleEditStyleAudio") then
+			if not fcp:performShortcut("MultiAngleEditStyleAudio") then
 				dialog.displayErrorMessage("We were unable to trigger the 'Cut/Switch Multicam Audio Only' Shortcut.\n\nPlease make sure this shortcut is allocated in the Command Editor.\n\nError Occured in cutAndSwitchMulticam().")
 				return "Failed"
 			end
 		end
 
 		if whichMode == "Video" then
-			if not fcp.app():performShortcut("MultiAngleEditStyleVideo") then
+			if not fcp:performShortcut("MultiAngleEditStyleVideo") then
 				dialog.displayErrorMessage("We were unable to trigger the 'Cut/Switch Multicam Video Only' Shortcut.\n\nPlease make sure this shortcut is allocated in the Command Editor.\n\nError Occured in cutAndSwitchMulticam().")
 				return "Failed"
 			end
 		end
 
 		if whichMode == "Both" then
-			if not fcp.app():performShortcut("MultiAngleEditStyleAudioVideo") then
+			if not fcp:performShortcut("MultiAngleEditStyleAudioVideo") then
 				dialog.displayErrorMessage("We were unable to trigger the 'Cut/Switch Multicam Audio and Video' Shortcut.\n\nPlease make sure this shortcut is allocated in the Command Editor.\n\nError Occured in cutAndSwitchMulticam().")
 				return "Failed"
 			end
 		end
 
-		if not fcp.app():performShortcut("CutSwitchAngle" .. tostring(string.format("%02d", whichAngle))) then
+		if not fcp:performShortcut("CutSwitchAngle" .. tostring(string.format("%02d", whichAngle))) then
 			dialog.displayErrorMessage("We were unable to trigger the 'Cut and Switch to Viewer Angle " .. tostring(whichAngle) .. "' Shortcut.\n\nPlease make sure this shortcut is allocated in the Command Editor.\n\nError Occured in cutAndSwitchMulticam().")
 			return "Failed"
 		end
@@ -5695,12 +5695,12 @@ end
 			clipboard.stopWatching()
 		end
 
-		if not fcp.app():performShortcut("Cut") then
+		if not fcp:performShortcut("Cut") then
 			dialog.displayErrorMessage("Failed to trigger the 'Cut' Shortcut.\n\nError occurred in moveToPlayhead().")
 			goto moveToPlayheadEnd
 		end
 
-		if not fcp.app():performShortcut("Paste") then
+		if not fcp:performShortcut("Paste") then
 			dialog.displayErrorMessage("Failed to trigger the 'Paste' Shortcut.\n\nError occurred in moveToPlayhead().")
 			goto moveToPlayheadEnd
 		end
@@ -5725,7 +5725,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Get Browser Persistent Playhead:
 		--------------------------------------------------------------------------------
-		local playhead = fcp.app():libraries():playhead()
+		local playhead = fcp:libraries():playhead()
 		if playhead:isShowing() then
 
 			--------------------------------------------------------------------------------
@@ -5845,7 +5845,7 @@ end
 		-- Set Custom Export Path (or Default to Desktop):
 		--------------------------------------------------------------------------------
 		local batchExportDestinationFolder = settings.get("fcpxHacks.batchExportDestinationFolder")
-		local NSNavLastRootDirectory = fcp.app():getPreference("NSNavLastRootDirectory")
+		local NSNavLastRootDirectory = fcp:getPreference("NSNavLastRootDirectory")
 		local exportPath = "~/Desktop"
 		if batchExportDestinationFolder ~= nil then
 			 if tools.doesDirectoryExist(batchExportDestinationFolder) then
@@ -5863,7 +5863,7 @@ end
 		local destinationPreset = settings.get("fcpxHacks.batchExportDestinationPreset")
 		if destinationPreset == nil then
 
-			destinationPreset = fcp.app():menuBar():findMenuUI("File", "Share", function(menuItem)
+			destinationPreset = fcp:menuBar():findMenuUI("File", "Share", function(menuItem)
 				return menuItem:attributeValue("AXMenuItemCmdChar") ~= nil
 			end):attributeValue("AXTitle")
 
@@ -5886,7 +5886,7 @@ end
 		--------------------------------------------------------------------------------
 		deleteAllHighlights()
 
-		local libraries = fcp.app():browser():libraries()
+		local libraries = fcp:browser():libraries()
 
 		if not libraries:isShowing() then
 			dialog.displayErrorMessage(i18n("batchExportEnableBrowser"))
@@ -5971,7 +5971,7 @@ end
 				--------------------------------------------------------------------------------
 				-- Wait for Export Dialog to open:
 				--------------------------------------------------------------------------------
-				local exportDialog = fcp.app():exportDialog()
+				local exportDialog = fcp:exportDialog()
 				if not just.doUntil(function() return exportDialog:isShowing() end) then
 					dialog.displayErrorMessage("Failed to open the 'Export' window.")
 					return -2
@@ -6018,7 +6018,7 @@ end
 		-- Trigger Export:
 		--------------------------------------------------------------------------------
 		function selectShare(destinationPreset)
-			return fcp.app():menuBar():selectMenu("File", "Share", function(menuItem)
+			return fcp:menuBar():selectMenu("File", "Share", function(menuItem)
 				if destinationPreset == nil then
 					return menuItem:attributeValue("AXMenuItemCmdChar") ~= nil
 				else
@@ -6057,7 +6057,7 @@ end
 		-- Variables:
 		--------------------------------------------------------------------------------
 		local ninjaPasteboardCopyError = false
-		local finalCutProClipboardUTI = fcp.app():getPasteboardUTI()
+		local finalCutProClipboardUTI = fcp:getPasteboardUTI()
 		local enableClipboardHistory = settings.get("fcpxHacks.enableClipboardHistory") or false
 
 		--------------------------------------------------------------------------------
@@ -6251,7 +6251,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Show Touch Bar at Top Centre of Timeline:
 		--------------------------------------------------------------------------------
-		local timeline = fcp.app():timeline()
+		local timeline = fcp:timeline()
 		if displayTouchBarLocation == "TimelineTopCentre" and timeline:isShowing() then
 			--------------------------------------------------------------------------------
 			-- Position Touch Bar to Top Centre of Final Cut Pro Timeline:
@@ -6319,7 +6319,7 @@ function finalCutProWindowWatcher()
 	fullscreenPlaybackWatcher:subscribe(windowfilter.windowCreated,(function(window, applicationName)
 		if applicationName == "Final Cut Pro" then
 			if window:title() == "" then
-				local fcpx = fcp.app():application()
+				local fcpx = fcp:application()
 				local fcpxElements = ax.applicationElement(fcpx)
 				if fcpxElements[1][1] ~= nil then
 					if fcpxElements[1][1]:attributeValue("AXIdentifier") == "_NS:523" then
@@ -6355,7 +6355,7 @@ function finalCutProWindowWatcher()
 	end), true)
 
 	-- Watch the command editor showing and hiding.
-	fcp.app():commandEditor():watch({
+	fcp:commandEditor():watch({
 		show = function(commandEditor)
 			--------------------------------------------------------------------------------
 			-- Disable Hotkeys:
@@ -6415,7 +6415,7 @@ function finalCutProWindowWatcher()
 	-- Final Cut Pro Window Not On Screen:
 	--------------------------------------------------------------------------------
 	finalCutProWindowFilter:subscribe(windowfilter.windowNotOnScreen, function()
-		if not fcp.app():isFrontmost() then
+		if not fcp:isFrontmost() then
 			finalCutProNotActive()
 		end
 	end, true)
@@ -6498,7 +6498,7 @@ end
 		timer.doAfter(0.0000000000001, function()
 			local lockTimelinePlayhead = settings.get("fcpxHacks.lockTimelinePlayhead") or false
 			if lockTimelinePlayhead then
-				fcp.app():timeline():lockPlayhead()
+				fcp:timeline():lockPlayhead()
 			end
 		end)
 
@@ -6522,7 +6522,7 @@ end
 		-- Update Current Language:
 		--------------------------------------------------------------------------------
 		timer.doAfter(0.0000000000001, function()
-			fcp.app():getCurrentLanguage(true)
+			fcp:getCurrentLanguage(true)
 		end)
 
 	end
@@ -6565,7 +6565,7 @@ end
 		--------------------------------------------------------------------------------
 		local lockTimelinePlayhead = settings.get("fcpxHacks.lockTimelinePlayhead") or false
 		if lockTimelinePlayhead then
-			fcp.app():timeline():unlockPlayhead()
+			fcp:timeline():unlockPlayhead()
 		end
 
 		--------------------------------------------------------------------------------
@@ -6620,8 +6620,8 @@ function finalCutProSettingsWatcher(files)
 		--------------------------------------------------------------------------------
 		-- Refresh Keyboard Shortcuts if Command Set Changed & Command Editor Closed:
 		--------------------------------------------------------------------------------
-    	if mod.lastCommandSet ~= fcp.app():getActiveCommandSetPath() then
-    		if not fcp.app():commandEditor():isShowing() then
+    	if mod.lastCommandSet ~= fcp:getActiveCommandSetPath() then
+    		if not fcp:commandEditor():isShowing() then
 	    		timer.doAfter(0.0000000000001, function() bindKeyboardShortcuts() end)
 			end
 		end
@@ -6665,7 +6665,7 @@ function fullscreenKeyboardWatcher()
 		--------------------------------------------------------------------------------
 		-- Define Final Cut Pro:
 		--------------------------------------------------------------------------------
-		local fcpx = fcp.app():application()
+		local fcpx = fcp:application()
 		local fcpxElements = ax.applicationElement(fcpx)
 
 		--------------------------------------------------------------------------------
@@ -6759,7 +6759,7 @@ function mediaImportWatcher()
 
 			debugMessage("Media Inserted.")
 
-			local mediaImport = fcp.app():mediaImport()
+			local mediaImport = fcp:mediaImport()
 			
 			if mediaImport:isShowing() then
 				-- Media Import was already open. Bail!
@@ -6772,20 +6772,20 @@ function mediaImportWatcher()
 			local currentApplication = application.frontmostApplication()
 			debugMessage("Currently using '"..currentApplication:name().."'")
 
-			local fcpxHidden = not fcp.app():isShowing()
+			local fcpxHidden = not fcp:isShowing()
 
 			mediaImportTimer = timer.doUntil(
 				function()
 					return stopMediaImportTimer
 				end,
 				function()
-					if not fcp.app():isRunning() then
+					if not fcp:isRunning() then
 						debugMessage("FCPX is not running. Stop watching.")
 						stopMediaImportTimer = true
 					else
 						if mediaImport:isShowing() then
 							mediaImport:hide()
-							if fcpxHidden then fcp.app():hide() end
+							if fcpxHidden then fcp:hide() end
 							currentApplication:activate()
 							debugMessage("Hid FCPX and returned to '"..currentApplication:name().."'.")
 							stopMediaImportTimer = true
@@ -6810,7 +6810,7 @@ end
 --------------------------------------------------------------------------------
 function scrollingTimelineWatcher()
 
-	local timeline = fcp.app():timeline()
+	local timeline = fcp:timeline()
 
 	--------------------------------------------------------------------------------
 	-- Key Press Down Watcher:
@@ -6839,7 +6839,7 @@ end
 		--------------------------------------------------------------------------------
 		-- Make sure the Command Editor and hacks console are closed:
 		--------------------------------------------------------------------------------
-		if fcp.app():commandEditor():isShowing() or hacksconsole.active then
+		if fcp:commandEditor():isShowing() or hacksconsole.active then
 			debugMessage("Spacebar pressed while other windows are visible.")
 			return "Stop"
 		end
@@ -6847,12 +6847,12 @@ end
 		--------------------------------------------------------------------------------
 		-- Don't activate scrollbar in fullscreen mode:
 		--------------------------------------------------------------------------------
-		if fcp.app():fullScreenWindow():isShowing() then
+		if fcp:fullScreenWindow():isShowing() then
 			debugMessage("Spacebar pressed in fullscreen mode whilst watching for scrolling timeline.")
 			return "Stop"
 		end
 
-		local timeline = fcp.app():timeline()
+		local timeline = fcp:timeline()
 
 		--------------------------------------------------------------------------------
 		-- Get Timeline Scroll Area:
@@ -6965,7 +6965,7 @@ function sharedXMLFileWatcher(files)
 				if host.localizedName() ~= editorName then
 
 					local xmlSharingPath = settings.get("fcpxHacks.xmlSharingPath")
-					sharedXMLNotification = notify.new(function() fcp.app():importXML(file) end)
+					sharedXMLNotification = notify.new(function() fcp:importXML(file) end)
 						:setIdImage(image.imageFromPath(fcpxhacks.iconPath))
 						:title("New XML Recieved")
 						:subTitle(file:sub(string.len(xmlSharingPath) + 1 + string.len(editorName) + 1, -8))
